Use 'python3' as Python executable in setup.sh (#3)
Overview
--------
Customers have reported that they can't run the `setup.sh` script of a freshly-created
Roboto action, because it runs Python as `python`, which they have to change to `python3`.

Reading PEP-394, it seems that as long as Python 3 is installed, the `python3` command will
be available. Reading StackOverflow, it seems like some Linux distros will only ship with
`python3` as a viable command, with extra options for people who want to symlink `python` to
`python3`.

For these reasons, I feel it's safe enough to change our default to `python3`.

Bonus
-------
Some README instructions on how to develop in this repo.

Testing
--------
* make install && make fmt && make lint > OK
* .venv/bin/cookiecutter --output-dir /tmp/demo_action . > created demo Python action and
  deployed it to my dev account. Action was invocable and did what the demo code does.
* Updated setup.sh for my custom action I'd previously created with `roboto actions init`.
  Re-build and re-deployment worked OK, and the action ran as expected.